requirements.txt
google-auth==2.1.1
google-auth-oauthlib==0.4.4
google-auth-httplib2==0.1.0
google-api-python-client==2.25.0
mysql-connector-python==8.0.28
pandas==1.3.3
pymongo==3.12.0
streamlit==1.16.0
streamlit-option-menu==1.1.2
